feat: resolve issue #8 — all 8 post-audit friction points
